Move stateful tests to public testing API (#2531)
* Move stateful tests to public testing API

* consolidate a bit
